[apps/desktop] move feed utilities
